move initial state from store to src/index.js
