+++ src/index.js
@@ -2,9 +2,27 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import { Provider } from 'react-redux';
+import { createStore } from 'redux';
 import App from './components/App';
-import store from './store';
+import combineReducers from './reducers/index';
 
+const state = {
+  books: [
+    {
+      id: Math.random(),
+      title: 'Limitless',
+      category: 'Learning',
+    },
+
+    {
+      id: Math.random(),
+      title: 'Scala fundamentals',
+      category: 'Learning',
+    },
+  ],
+};
+
+const store = createStore(combineReducers, state);
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
